tls changes and prepare for implementing oci

+++ rays/env.go
@@ -108,7 +108,7 @@ func updateProjects(updateRollbacks bool) {//we wont print anything if no update
 			process := getProcessFromBranch(deployment.Branch, project)
 			if (process == nil) {continue}
 
-			//if we're rolled back and not on the faulty version
+			//if we're rolled back and the newest version is the faulty version
 			if strings.HasPrefix(process.Hash, "rollback:") && strings.Replace(process.Hash, "rollback:", "", 1) == branches[deployment.Branch] && !updateRollbacks {
 				continue
 			}
@@ -140,7 +140,7 @@ func finishProcess(logFile logFile, process process, project project, branch str
 		logFile.Success = true
 		go triggerEvent("newProcess", process)
 	} else {
-		rlog.Notify(project.Name + ", branch " + branch + " was not sucessfully deployed!", "err")
+		rlog.Notify(project.Name + ", branch " + branch + " was not successfully deployed!", "err")
 		logFile.Success = false
 		go triggerEvent("processError", process)
 		go taskAutofix(process)
@@ -275,7 +275,7 @@ func deployLocalProcess(configPath string, dir string, project *project, swapfun
 		}
 		cmd.Env = append(cmd.Env, "RAY_DEPLOYMENT=" + process.Branch)
 
-		if (step.Type == "deploy") {
+		if step.Type == "deploy" {
 			(*swapfunction)()
 		}
 
@@ -350,11 +350,19 @@ func setupLocalProject(project *project, host string, hardCommit string) []proce
 			oldprocesses = append(oldprocesses, prc)
 		}
 	}
+	rm := func() {
+		for _, proc := range oldprocesses {
+			if (!proc.Ghost) {
+				proc.remove()
+			}
+		}
+	}
 
 	var deployments = project.Deployments
 	deployments = append(deployments, deployment{
 		Type: "prod",
 		Branch: "prod",
+		OCISrc: project.ProdOCISrc,
 	})
 
 	branchHashes := getBranches(project.Src)
@@ -365,60 +373,57 @@ func setupLocalProject(project *project, host string, hardCommit string) []proce
 		dir := filepath.Join(rdata.RayEnv, procId)
 		os.Mkdir(dir, 0600)
 
-		_cmd := []string{"clone", project.Src}
-		if (deployment.Type != "prod") {
-			_cmd = append(_cmd, "-b")
-			_cmd = append(_cmd, deployment.Branch)
-		}
-		
-		gitOutput := strings.Builder{}
-		cmd := exec.Command("git", _cmd...)
-		cmd.Dir = dir
-		cmd.Stdout = &gitOutput
-		cmd.Stderr = &gitOutput
-		
-		if err := cmd.Run(); err != nil {
-			rlog.Println(_cmd)
-			rlog.Println(gitOutput.String())
-			rlog.Notify("Git cloning error: " + err.Error(), "err")
-			continue //this wont fire any monitoring stuff and just silently fail (with the exception of the above log), could def be improved
-		}
-	
-		content, err := os.ReadDir(dir)
-		if err != nil {
-			rlog.Fatal(err)
-		}
-
-		if hardCommit != "" {
-			cmd := exec.Command("git", "reset", "--hard", hardCommit)
-			cmd.Dir = path.Join(dir, content[0].Name())
+		if deployment.OCISrc == "" {
+			_cmd := []string{"clone", project.Src}
+			if (deployment.Type != "prod") {
+				_cmd = append(_cmd, "-b")
+				_cmd = append(_cmd, deployment.Branch)
+			}
+			
+			gitOutput := strings.Builder{}
+			cmd := exec.Command("git", _cmd...)
+			cmd.Dir = dir
+			cmd.Stdout = &gitOutput
+			cmd.Stderr = &gitOutput
+			
 			if err := cmd.Run(); err != nil {
-				rlog.Println(cmd.Args)
-				rlog.Notify("Git hard commit resetting error: " + err.Error(), "err")
+				rlog.Println(_cmd)
+				rlog.Println(gitOutput.String())
+				rlog.Notify("Git cloning error: " + err.Error(), "err")
 				continue //this wont fire any monitoring stuff and just silently fail (with the exception of the above log), could def be improved
 			}
-		}
+		
+			content, err := os.ReadDir(dir)
+			if err != nil {
+				rlog.Fatal(err)
+			}
 
-		os.Mkdir(path.Join(dir, "logs"), 0600)//Making sure to do this after we've cloned the repo
-	
-		projectConfig := filepath.Join(dir, content[0].Name(), "ray.config.json") //the existance of this file is checked later dw
-		rm := func() {
-			for _, proc := range oldprocesses {
-				if (!proc.Ghost) {
-					proc.remove()
+			if hardCommit != "" {
+				cmd := exec.Command("git", "reset", "--hard", hardCommit)
+				cmd.Dir = path.Join(dir, content[0].Name())
+				if err := cmd.Run(); err != nil {
+					rlog.Println(cmd.Args)
+					rlog.Notify("Git hard commit resetting error: " + err.Error(), "err")
+					continue //this wont fire any monitoring stuff and just silently fail (with the exception of the above log), could def be improved
 				}
 			}
-		}
-	
-		branch := deployment.Branch
+			os.Mkdir(path.Join(dir, "logs"), 0600) //Making sure to do this after we've cloned the repo and called ReadDir
 
-		branchHash := ""
-		if hardCommit != "" {
-			branchHash = "rollback:" + branchHashes[branch] //the faulty version
-		} else if (branchHashes != nil && branchHashes[branch] != "") {
-			branchHash = branchHashes[branch]
+			projectConfig := filepath.Join(dir, content[0].Name(), "ray.config.json") //the existance of this file is checked later dw
+			branch := deployment.Branch
+			branchHash := ""
+
+			if hardCommit != "" {
+				branchHash = "rollback:" + branchHashes[branch] //the faulty version
+			} else if (branchHashes != nil && branchHashes[branch] != "") {
+				branchHash = branchHashes[branch]
+			}
+
+			go deployLocalProcess(projectConfig, filepath.Join(dir, content[0].Name()), project, &rm, branch, branchHash, path.Join(dir, "logs"), dir, procId, host)
+		} else {
+			os.Mkdir(path.Join(dir, "logs"), 0600)
+			//todo: implement
 		}
-		go deployLocalProcess(projectConfig, filepath.Join(dir, content[0].Name()), project, &rm, branch, branchHash, path.Join(dir, "logs"), dir, procId, host)
 	}
 
 	var newProcesses []process
